requirements.txt
fastapi[all]
uvicorn
pydantic
sqlalchemy
python-dotenv
python-jose[cryptography]
